fix(ci): reduce complexity hotspots for waste sync
