+++ packages/core/src/waste-management-operations-contract.ts
@@ -61,12 +61,6 @@ export type WasteManagementResetJobInput = {
   readonly confirmationToken: string;
 };
 
-export type WasteManagementSyncMainserverJobInput = {
-  readonly operation: 'sync-mainserver';
-  readonly keycloakSubject?: string;
-  readonly activeOrganizationId?: string;
-};
-
 export type WasteManagementSyncWasteTypesJobInput = {
   readonly operation: 'sync-waste-types';
   readonly keycloakSubject?: string;
@@ -79,7 +73,11 @@ export type WasteManagementJobInput =
   | WasteManagementImportJobInput
   | WasteManagementSeedJobInput
   | WasteManagementResetJobInput
-  | WasteManagementSyncMainserverJobInput
+  | {
+      readonly operation: 'sync-mainserver';
+      readonly keycloakSubject?: string;
+      readonly activeOrganizationId?: string;
+    }
   | WasteManagementSyncWasteTypesJobInput;
 
 export const wasteManagementOperationsContract = {
