+++ packages/plugin-sdk/src/index.ts
@@ -138,7 +138,6 @@ export type {
   WasteManagementMasterDataOverview,
   WasteManagementResetJobInput,
   WasteManagementSeedJobInput,
-  WasteManagementSyncMainserverJobInput,
   WasteManagementSyncWasteTypesJobInput,
   WasteManagementSettingsInterfaceOption,
   WasteManagementSettingsRecord,
