+++ packages/plugin-sdk/src/public-api.ts
@@ -40,7 +40,6 @@ export type {
   WasteManagementMasterDataOverview,
   WasteManagementResetJobInput,
   WasteManagementSeedJobInput,
-  WasteManagementSyncMainserverJobInput,
   WasteManagementSyncWasteTypesJobInput,
   WasteManagementSettingsInterfaceOption,
   WasteManagementSettingsRecord,
