+++ packages/plugin-waste-management/src/waste-management.page.tsx
@@ -42,16 +42,11 @@ const updateSearch = (
   });
 };
 
-export const WasteManagementPage = () => {
-  const pt = usePluginTranslation('wasteManagement');
-  const navigate = useNavigate();
-  const rawSearch = useSearch({ strict: false });
-  const search = normalizeWasteManagementSearchParams(rawSearch as Record<string, unknown>);
-  const uiAccess = useWasteManagementUiAccess(search.tab);
-  const [calendarWebUrl, setCalendarWebUrl] = useState<string | null>(null);
-  const [syncRunning, setSyncRunning] = useState(false);
-  const [statusMessage, setStatusMessage] = useState<StatusMessage | null>(null);
-
+const useWasteManagementVisibleTabRedirect = (
+  navigate: ReturnType<typeof useNavigate>,
+  search: WasteManagementSearchParams,
+  uiAccess: ReturnType<typeof useWasteManagementUiAccess>
+) => {
   useEffect(() => {
     if (!uiAccess.isResolved || uiAccess.visibleTabIds.includes(search.tab)) {
       return;
@@ -73,6 +68,10 @@ export const WasteManagementPage = () => {
       replace: true,
     });
   }, [navigate, search, uiAccess.isResolved, uiAccess.visibleTabIds]);
+};
+
+const useWasteManagementCalendarWebUrl = (uiAccess: ReturnType<typeof useWasteManagementUiAccess>) => {
+  const [calendarWebUrl, setCalendarWebUrl] = useState<string | null>(null);
 
   useEffect(() => {
     if (!uiAccess.isResolved || !uiAccess.canAccessSettings) {
@@ -100,6 +99,67 @@ export const WasteManagementPage = () => {
     };
   }, [uiAccess.canAccessSettings, uiAccess.isResolved]);
 
+  return calendarWebUrl;
+};
+
+const useWasteManagementSyncAction = (pt: ReturnType<typeof usePluginTranslation>) => {
+  const [syncRunning, setSyncRunning] = useState(false);
+  const [statusMessage, setStatusMessage] = useState<StatusMessage | null>(null);
+
+  const startSync = async () => {
+    setStatusMessage(null);
+    setSyncRunning(true);
+    try {
+      await startWasteManagementMainserverSync({});
+      setStatusMessage({
+        kind: 'success',
+        text: pt('tools.sync.startSuccess'),
+      });
+    } catch {
+      setStatusMessage({
+        kind: 'error',
+        text: pt('tools.sync.startError'),
+      });
+    } finally {
+      setSyncRunning(false);
+    }
+  };
+
+  return {
+    syncRunning,
+    statusMessage,
+    startSync,
+  };
+};
+
+const WasteManagementSyncAction = ({
+  canRunMainserverSync,
+  disabled,
+  onClick,
+  pt,
+}: Readonly<{
+  canRunMainserverSync: boolean;
+  disabled: boolean;
+  onClick: () => Promise<void>;
+  pt: ReturnType<typeof usePluginTranslation>;
+}>) =>
+  canRunMainserverSync ? (
+    <Button type="button" disabled={disabled} onClick={() => void onClick()}>
+      {pt('tools.sync.actionLabel')}
+    </Button>
+  ) : null;
+
+export const WasteManagementPage = () => {
+  const pt = usePluginTranslation('wasteManagement');
+  const navigate = useNavigate();
+  const rawSearch = useSearch({ strict: false });
+  const search = normalizeWasteManagementSearchParams(rawSearch as Record<string, unknown>);
+  const uiAccess = useWasteManagementUiAccess(search.tab);
+  const calendarWebUrl = useWasteManagementCalendarWebUrl(uiAccess);
+  const { syncRunning, statusMessage, startSync } = useWasteManagementSyncAction(pt);
+
+  useWasteManagementVisibleTabRedirect(navigate, search, uiAccess);
+
   return (
     <StudioOverviewPageTemplate
       title={pt('page.title')}
@@ -112,32 +172,12 @@ export const WasteManagementPage = () => {
         />
       }
       primaryAction={
-        uiAccess.canRunMainserverSync ? (
-          <Button
-            type="button"
-            disabled={syncRunning}
-            onClick={async () => {
-              setStatusMessage(null);
-              setSyncRunning(true);
-              try {
-                await startWasteManagementMainserverSync({});
-                setStatusMessage({
-                  kind: 'success',
-                  text: pt('tools.sync.startSuccess'),
-                });
-              } catch {
-                setStatusMessage({
-                  kind: 'error',
-                  text: pt('tools.sync.startError'),
-                });
-              } finally {
-                setSyncRunning(false);
-              }
-            }}
-          >
-            {pt('tools.sync.actionLabel')}
-          </Button>
-        ) : null
+        <WasteManagementSyncAction
+          canRunMainserverSync={uiAccess.canRunMainserverSync}
+          disabled={syncRunning}
+          onClick={startSync}
+          pt={pt}
+        />
       }
       toolbar={statusMessage ? <StatusNotice message={statusMessage} /> : null}
     >
